aider: Fix: Removed <a> tag from within the <Link> component.
# Aider chat conversation:

USER: Error: Invalid <Link> with <a> child. Please remove <a> or use <Link legacyBehavior>.
Learn more: https://nextjs.org/docs/messages/invalid-new-link-with-extra-anchor

ASSISTANT: To resolve the error, we need to remove the `<a>` tag from within the `<Link>` component. Here's the diff for the necessary changes:

```diff
--- app/page.tsx
+++ app/page.tsx
@@ ... @@
 export default function HomePage() {

   return (
     <div className="bg-orange-500 h-screen flex flex-col items-center justify-center text-white">
       <h1>Do you ha(l)ve Bitcoin? We should party.</h1>
-      <Link href="/bitcoin">
-        <a>
-          <Button className="border-black border-2 w-1/10 h-3/20">
+      <Link href="/bitcoin" passHref>
+          <Button as="a" className="border-black border-2 w-1/10 h-3/20">
               PRESS ME PLEASE
-          </Button>
-        </a>
+          </Button>
       </Link>
     </div>
   )
 }
```

This change uses the `passHref` prop on the `Link` component to pass the `href` to the `Button` component, which should be able to accept it as an anchor tag if it's designed to do so. If the `Button` component is not designed to accept `href` as a prop, you would need to modify the `Button` component to handle this prop and render an `<a>` tag accordingly. If you need assistance with that, please provide the code for the `Button` component.