aider: Modify image width based on screen size.
# Aider chat conversation:

USER: how can i edit the image to be different sizes based on the screen width/height. my `tailwind.config.js` file has a screens value in the theme. Can you modify it so that the image takes up 1/2 the width if the screen size is larger than the md value?

ASSISTANT: To make the image take up half the width of the screen on medium-sized screens and larger, you can use Tailwind CSS's responsive utilities. You'll need to apply conditional classes based on the screen size. Here's how you can modify the `pages/intro.tsx` file to achieve this:

```diff
--- pages/intro.tsx
+++ pages/intro.tsx
@@ ... @@
 <div className="h-1/2 w-full flex justify-center items-center">
-<div className="relative w-1/2 h-1/2">
+<div className="relative w-full h-1/2 md:w-1/2">
   <img
         src="/images/notreal.jpeg"
         alt="not a real bitcoiner"
   />
 </div>
 </div>
```

This change applies the `w-full` class to make the image take the full width on smaller screens, and then applies the `md:w-1/2` class to make the image take up half the width on medium-sized screens (`md`) and larger. The `h-1/2` class is kept to maintain the height at half of the parent container's height.